// lib/types/types.ts
  export type DirectusFiles = {
    charset?: string | null;
    description?: string | null;
    duration?: number | null;
    embed?: string | null;
    filename_disk?: string | null;
    filename_download: string;
    filesize?: number | null;
    focal_point_x?: number | null;
    focal_point_y?: number | null;
    folder?: (string & DirectusFolders) | null;
    height?: number | null;
    id: string;
    location?: string | null;
    metadata?: unknown | null;
    modified_by?: (string & DirectusUsers) | null;
    modified_on: string;
    storage: string;
    tags?: unknown | null;
    title?: string | null;
    type?: string | null;
    uploaded_by?: (string & DirectusUsers) | null;
    uploaded_on: string;
    width?: number | null;
  };
// ...
  export type DirectusFolders = {
    id: string;
    name: string;
    parent?: (string & DirectusFolders) | null;
  };
// ...
  export type DirectusUsers = {
    appearance?: string | null;
    auth_data?: unknown | null;
    avatar?: (string & DirectusFiles) | null;
    description?: string | null;
    email?: string | null;
    email_notifications?: boolean | null;
    external_identifier?: string | null;
    first_name?: string | null;
    id: string;
    language?: string | null;
    last_access?: string | null;
    last_name?: string | null;
    last_page?: string | null;
    location?: string | null;
    password?: string | null;
    provider: string;
    role?: (string & DirectusRoles) | null;
    status: string;
    tags?: unknown | null;
    tfa_secret?: string | null;
    theme_dark?: string | null;
    theme_dark_overrides?: unknown | null;
    theme_light?: string | null;
    theme_light_overrides?: unknown | null;
    title?: string | null;
    token?: string | null;
  };
// ...
  export type Products = {
    date_created?: string | null;
    description?: string | null;
    id: number;
    image?: (string & DirectusFiles) | null;
    name?: string | null;
    price?: number | null;
    stock_quantity?: number | null;
    quantity: number;
  };